+++ mod_form.php
@@ -287,52 +287,6 @@ class mod_mumie_mod_form extends moodleform_mod {
     }
 
     /**
-     * Provide option to mark an activity automatically as completed once a passing grade was archived
-     *
-     * This function is copied from mod_quiz version 2018051400
-     * @return array containing the name of the mform group that has been added to the form
-     */
-    public function add_completion_rules(): array {
-        $mform = $this->_form;
-        $items = [];
-
-        $group = [];
-        $completionpasselement = $this->get_completion_rule_element_name('completionpass');
-        $group[] = $mform->createElement(
-            'advcheckbox',
-            $completionpasselement,
-            null,
-            get_string('completionpass', 'mumie'),
-            ['group' => 'cpass']
-        );
-        $completionusegradeelement = $this->get_completion_rule_element_name('completionusegrade');
-        $mform->disabledIf($completionpasselement, $completionusegradeelement, 'notchecked');
-        $completionpassgroupelement = $this->get_completion_rule_element_name('completionpassgroup');
-        $mform->addGroup($group, $completionpassgroupelement, get_string('completionpass', 'mumie'), ' &nbsp; ', false);
-        $mform->addHelpButton($completionpassgroupelement, 'completionpass', 'mumie');
-        $items[] = $completionpassgroupelement;
-        return $items;
-    }
-
-    /**
-     * Get the completion rule's element name.
-     *
-     * Conditionally add suffix for Moodle >= 4.3.
-     *
-     * @param string $rawname The raw name of the completion rule.
-     * @return string The properly suffixed element name.
-     */
-    private function get_completion_rule_element_name($rawname): string {
-        global $CFG;
-        if ($CFG->branch < 403) {
-            $suffix = '';
-        } else {
-            $suffix = $this->get_suffix();
-        }
-        return $rawname . $suffix;
-    }
-
-    /**
      * Disable all options for grades if the user has chosen to link a course instead of a problem.
      */
     private function disable_grade_rules(): void {
@@ -617,18 +571,6 @@ class mod_mumie_mod_form extends moodleform_mod {
             }
         }
     }
-
-    /**
-     * Called during validation. Indicates whether a module-specific completion rule is selected.
-     *
-     * @param array $data Input data (not yet validated)
-     * @return bool True if one or more rules is enabled, false if none are.
-     */
-    public function completion_rule_enabled($data): bool {
-        $completionpasselement = $this->get_completion_rule_element_name('completionpass');
-        return !empty($data[$completionpasselement]);
-    }
-
 
     /**
      * The decision regarding gradepools is final. We need to know whether we should disable the selection boxes.